Added support for outputting multiple files
